# perf: improve collections sorting query